Rewrite the renaming logic with hooks
