+++ packages/block-editor/src/components/block-rename/index.js
@@ -1,0 +1,3 @@
+export { default as BlockRenameMenuItem } from './menu-item';
+export { default as BlockRenameModal } from './modal';
+export { default as useBlockRename } from './use-block-rename';
