+++ packages/block-editor/src/components/block-rename/style.scss
